Add support for sample loop points
Exporting samples always includes the loop points and the loop enabled ("on",
"off") as stored in the bank.

When using `-s`, the sample is added to the bank, the loop points are set as in
the file and the loop enabled is set as in the file. If the sample has no loop
information ("smpl" chunk is missing), the loop points are set to 0 and the last
frame and the loop enabled is set to "off".

If the sample is imported with `-S` (force loop), the loop points are set to 0
and the last frame and the loop enabled is set to "on".