Add option to add an SFZ instrument to an emu3 bank
In order to be permissive with custom opcodes and simplify the code,
all grammar-allowed opcodes are parsed and stored in hashtables (GLib)
and later processed. This keeps emu3bm from being a full SFZ checker.

libsamplerate is now a dependency to better work with other sample rates.

Another improvement is that any sample imported is converted to 44100 Hz
if the original sampling rate is greater than that to be compatible with
the hardware.