Remove `-S` option
This option was forcing the loop to be set at the whole sample. This
had the inconvenience of setting the loop points outside the hardware
limits (0 and frames - 1 are both outside the allowed range).

Besides, since a few commits back, appending a sample honours its loop
points and fallbacks to the whole allowed sample if no loop points are
found.