Fix `Mono Start Zero!!!001` error
The sampler needs the first 2 and last 2 samples to be set to 0. Without
this, the sampler will complain with a "Mono Start Zero!!!001" error message.
As indicated in the manual, running the sample integrity in the digital tools
menu would fix the error and it'll fix it by doing this.

In previous versions of emu3bm, the pairs of 0 samples were appended
automatically but this broke SDS compatibility frame count.